% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFrequencies.R
\name{computeFrequencies}
\alias{computeFrequencies}
\title{Normalized Frequencies.}
\usage{
computeFrequencies(corpus, sentimentWord, shifterWord, clusterSize = 1)
}
\arguments{
\item{corpus}{corpus with column \code{docID}, \code{regID}, and \code{texts}. 
docID indicate a unique ID for the document. regID is used for aggregation of document for the regression. It must match with the specific regID of the response variable in  \code{\link{fitGWP}}.
texts are the textual data.}

\item{sentimentWord}{Vector of words used for computing the sentiment.}

\item{shifterWord}{Matrix with element \code{x} and \code{y}. x is a vector of valence shifting words while y is the modifier values.}

\item{clusterSize}{Scalar indicating the window in which valance shifting words have an influence.}
}
\value{
A list with the following elements:
        \itemize{
        \item docID: unique document ID.
        \item regID: regression ID.
        \item loc: location of the sentiment word within the texts.
        \item word: sentiment word. 
        \item shift: shigt coefficient modifying the sentiment word.
        \item NormalizedFrequency: frequency of word normalized by number of token in each text
        \item NormalizedFrequencyPerRegID: Normalzied Frequency normalized by the number of documents per each regression ID.
        }
}
\description{
Generate the Normalized Frequency table from textual data input.
}
\examples{
# Load example data
data("corpus",  package = "GWP")

# Setup the lexicons
sentimentWord <- sentometrics::list_lexicons$LM_en$x
shifterWord <- sentometrics::list_valence_shifters$en[, c("x", "y")]

# Generate the frequency data
frequencies <- computeFrequencies(corpus, sentimentWord, shifterWord, clusterSize = 5)
}
