% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWP.R
\docType{data}
\name{corpus}
\alias{corpus}
\title{Texts (not) relevant to the U.S. economy}
\format{A data frame with 4145 rows and 3 variables:
\describe{
\item{docID}{Unique document ID}
  \item{regID}{Date in format yyyy-mm}
  \item{texts}{Texts in \code{character} format.}
}}
\source{
\href{https://www.figure-eight.com/data-for-everyone/}{Economic News Article Tone and Relevance}. Retrieved
November 1, 2017.
}
\usage{
data("corpus")
}
\description{
A collection of texts annotated by humans in terms of relevance to the U.S. economy or not. The texts come from two major
journals in the U.S. (The Wall Street Journal and The Washington Post) and cover 4145 documents between 1995 and 2014. It
contains following information:
}
\keyword{datasets}
