% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWP.R
\docType{data}
\name{epu}
\alias{epu}
\title{EPU Index}
\format{A data frame with 403 rows and 2 variables:
\describe{
  \item{regID}{Date in format yyyy-mm}
  \item{y}{The EPU level}
}}
\source{
\url{https://www.policyuncertainty.com/us_monthly.html}
}
\usage{
data("epu")
}
\description{
A dataset containing the monthly level of the US Economic Policy Uncertainty (EPU) Index.
}
\keyword{datasets}
