% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictGWP}
\alias{predictGWP}
\title{Predict function for new data.}
\usage{
predictGWP(frequencies, model)
}
\arguments{
\item{frequencies}{frequencies data.table generated using the function \code{\link{computeFrequencies}}.}

\item{model}{Score model generated with \code{\link{fitGWP}}.}
}
\value{
A data.table with the following element:
        \itemize{
        \item regID: regression ID.
        \item SentimentScores: predicted sentiment scores.
        }
}
\description{
Ouput sentiment value given a fitted model and frequency table.
}
\examples{
# Load example data
data("corpus",  package = "GWP")
data("vix",  package = "GWP")
# Setup the lexicons
sentimentWord <- sentometrics::list_lexicons$LM_en$x
shifterWord <- sentometrics::list_valence_shifters$en[, c("x", "y")]

# Generate the frequency data
frequencies <- computeFrequencies(corpus[1:991, ], sentimentWord, shifterWord, clusterSize = 5)

# Calibrate the lexicon
res <- fitGWP(frequencies = frequencies, responseData = vix)

newFrequencies <- computeFrequencies(corpus[-1:-991, ], sentimentWord, shifterWord, clusterSize = 5)
pred <- predictGWP(newFrequencies, res)
}
