% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWP.R
\docType{data}
\name{vix}
\alias{vix}
\title{VIX Index}
\format{A data frame with 345 rows and 2 variables:
\describe{
  \item{regID}{Date in format yyyy-mm}
  \item{y}{The VIX level}
}}
\source{
\url{https://fred.stlouisfed.org/series/VIXCLS}
}
\usage{
data("vix")
}
\description{
A dataset containing the monthly level of the VIX Index.
}
\keyword{datasets}
